requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud==0.34.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.3.0
